Added call to fillWAMFields in DATA_INIT
To do this, we needed to expose the wwg,uug,vvg,n2g, etc.. fields in the
module_ATM_GRID_COMP.F90

To make this possible, 'use get_variables_for_wam_ipe_coupling' is added to this module
This also required that the make system can expose this dependency. This was possible only
if the get_variables_for_wam_ipe_coupling module was separated from get_w_z.f and placed
in its own .F90 file under NEMS/src/atmos/share/ . This move also required an update of the
makefile under the share/ directory. Currently, the make system indicates that the variables
are exposed, but there is currently a type mismatch in the call to fillWAMFields that needs to
be worked out. we are almost there