Added script to be used for quickly running model or making plots [#160]
This script has command line options that can be used to either run a
model forecast or generate plots.

The JOBNAME in the config file is set equal to the hash id of the
version that is being run.

Plots are directed to a common directory on theia that can be monitored
on a personal workstation. Underneath that common directory, the plots
are sorted into subdirectories according to the hash id