Fixing the build system on Theia [#196]
Previously we had used a hard path to the library directory of a user
in the meso group on Theia. That directory does not exist anymore, and
no soft-link has been provided, so I took the opportunity to abstract
the lib/incmod pathing into the module load, as is done with FV3.

There was also a hardpath in NEMS/src/atmos/gsm/phys/makefile for the
sigio incmod folder; that has also been abstracted to the module load.

This should be more flexible overall... in theory, we could use the
configure.nems.Theia.intel[_gsm] files on other systems if we
accordingly updated the modulefiles and the selection of the conf
files in NEMSAppBuilder, but that is probably unnecessary for now.