Fix remaining issues for #183
This commit adds a check to the exglobal script which correctly links
the model output from IPE for the coupled (ipe_grid_*_params.$IPEDATE)
and standalone (ipe_grid_*_params.iter_$ELAPSED_SECONDS). Additionally
the scripts now dynamically select the name of the standalone IPE
executable based on the current $MACHINE and $PARALLELISM if designated.