Issue #210
Branch for Houjun's changes which are now intended for development.

Holds wl=0 during idea_phys radiation blending to apply only IDEA
radiation in idea_phys, and then GFS radiation is treated separately
in gbphys, performing the blending process again after the call to
dcyc2, but setting wh=0 instead.

This seems to improve the tropospheric bias and have the least impact
on upper atmospheric temperatures. Further research is needed to find
a cohesive unified solution, but this is a good band-aid for now.